refactor: - reorganize top-rated model imports - remove reviews dto local it was not used - rearrange authentication preferences
