+++ data/src/main/java/com/london/data/local/model/home/topRated/TopRatedLocal.kt
@@ -1,4 +1,4 @@
-package com.london.data.local.model.home
+package com.london.data.local.model.home.topRated
 
 import androidx.room.Entity
 import com.london.domain.entity.recent.MediaType
