+++ data/src/test/java/com/london/data/repository/TopRatedTvSeriesRepositoryImplTest.kt
@@ -1,7 +1,7 @@
 package com.london.data.repository
 
 import com.google.common.truth.Truth.assertThat
-import com.london.data.local.model.home.TopRatedLocal
+import com.london.data.local.model.home.topRated.TopRatedLocal
 import com.london.data.local.source.home.HomeLocalDataSource
 import com.london.data.mapper.toprated.toEntity
 import com.london.data.remote.model.ApiResponse
